requirements.txt
apt-xapian-index==0.49
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.7.2
attrs==19.3.0
Automat==0.8.0
autopep8==2.0.4
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==23.2.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.7
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.12.3
future==0.18.2
gpg==1.13.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.1.2
pandas==2.0.3
paramiko==2.6.0
pbr==5.11.1
pexpect==4.6.0
Pillow==7.0.0
platformdirs==3.10.0
progressbar==2.5
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.11.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-gnupg==0.4.5
pytz==2023.3.post1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sos==4.4
sqlparse==0.4.4
ssh-import-id==5.10
stevedore==5.1.0
systemd-python==234
texttable==1.6.2
tomli==2.0.1
Twisted==18.9.0
typing-extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
Ubuntu-Make==18.9
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.24.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.3
websocket-client==0.53.0
xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
